run/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bitarray==2.7.3
boto3==1.28.9
botocore==1.31.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
et-xmlfile==1.1.0
filelock==3.12.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
fsspec==2023.4.0
grpcio==1.59.3
grpcio-tools==1.59.3
gunicorn==20.1.0
hdbcli==2.18.27
huggingface-hub==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parsel==1.8.1
pdfminer.six==20221105
pdfplumber==0.9.0
pdfreader==0.1.12
pdfrw==0.4
Pillow==9.5.0
Protego==0.2.1
protobuf==4.25.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
queuelib==1.6.2
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
s3transfer==0.6.1
scikit-learn==1.3.2
scipy==1.11.4
Scrapy==2.9.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.10
stripe==5.4.0
threadpoolctl==3.2.0
tldextract==3.4.1
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
Wand==0.6.11
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.8.2
zope.interface==6.0
